backend/requirements.txt
Django==5.2.5
djangorestframework==3.15.2
django-cors-headers==4.4.0
Pillow==10.4.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0
channels==4.0.0
channels-redis==4.2.0
asgiref==3.7.2

# Email Services
sendgrid==6.11.0
email-validator==2.2.0

# Additional utilities
requests==2.32.3
beautifulsoup4==4.12.3
